+++ src/shared/components/label-component/index.ts
@@ -1,0 +1,1 @@
+export * from './label-component';
