+++ src/shared/components/label-component/label-component.module.scss
