--- src/shared/components/status-label/index.ts
@@ -1,1 +1,0 @@
-export * from './status-label';
